--- posu/units/si/si.hpp
@@ -1,13 +1,0 @@
-#ifndef POSU_UNITS_SI_SI_HPP
-#define POSU_UNITS_SI_SI_HPP
-
-#include "posu/units/si/amount_of_substance.hpp"
-#include "posu/units/si/electric_current.hpp"
-#include "posu/units/si/length.hpp"
-#include "posu/units/si/luminous_intensity.hpp"
-#include "posu/units/si/mass.hpp"
-#include "posu/units/si/speed.hpp"
-#include "posu/units/si/thermodynamic_temperature.hpp"
-#include "posu/units/si/time.hpp"
-
-#endif // #ifndef POSU_UNITS_SI_SI_HPP
