+++ posu/units/system/si/mass.hpp
@@ -90,6 +90,6 @@ namespace posu::units::si {
 
 } // namespace posu::units::si
 
-#include "posu/units/si/ipp/mass.ipp"
+#include "posu/units/system/si/ipp/mass.ipp"
 
 #endif // #ifndef POSU_UNITS_SI_MASS_HPP
