+++ posu/units/system/si/speed.hpp
@@ -2,8 +2,8 @@
 #define POSU_UNITS_SI_SPEED_HPP
 
 #include "posu/units/base_unit.hpp"
-#include "posu/units/si/length.hpp"
-#include "posu/units/si/time.hpp"
+#include "posu/units/system/si/length.hpp"
+#include "posu/units/system/si/time.hpp"
 #include "posu/units/system/velocity.hpp"
 
 namespace posu::units::si {
@@ -110,6 +110,6 @@ namespace posu::units::si {
 
 } // namespace posu::units::si
 
-#include "posu/units/si/ipp/speed.ipp"
+#include "posu/units/system/si/ipp/speed.ipp"
 
 #endif // #ifndef POSU_UNITS_SI_SPEED_HPP
