+++ test/units/system/si/test_amount_of_substance.cpp
@@ -1,4 +1,4 @@
-#include "posu/units/si/amount_of_substance.hpp"
+#include "posu/units/system/si/amount_of_substance.hpp"
 
 #include <catch2/catch.hpp>
