+++ test/units/system/si/test_electric_current.cpp
@@ -1,4 +1,4 @@
-#include "posu/units/si/electric_current.hpp"
+#include "posu/units/system/si/electric_current.hpp"
 
 #include <catch2/catch.hpp>
