+++ test/units/system/si/test_mass.cpp
@@ -1,4 +1,4 @@
-#include "posu/units/si/mass.hpp"
+#include "posu/units/system/si/mass.hpp"
 
 #include <catch2/catch.hpp>
