+++ test/units/system/si/test_speed.cpp
@@ -1,4 +1,4 @@
-#include "posu/units/si/speed.hpp"
+#include "posu/units/system/si/speed.hpp"
 
 #include <catch2/catch.hpp>
